fixed warning in ageoffile
git-svn-id: https://stage.maemo.org/svn/maemo/projects/haf/trunk@6050 50eb543b-83fa-0310-9bb7-f7cbbee4ab39

+++ osso-af-utils/src/ageoffile.c
@@ -14,6 +14,6 @@ int main(int argc, char* argv[])
                 perror("lstat");
                 return 1;
         }
-        printf("%u\n", buf.st_mtime);
+        printf("%u\n", (unsigned)buf.st_mtime);
         return 0;
 }
